Remove deprecated user object fields
Wrapping a Maybe around the userLang field (which was done in a previous commit) was just the tip of the iceberg. There are many other fields that are now nullable that weren't previously and will cause JSON parsing errors. We *could* just wrap Maybes around all of those fields, but as those fields are always going to be null from now on, it's probably better to just remove them altogether.

https://twittercommunity.com/t/upcoming-changes-to-user-object-and-get-users-suggestions-endpoints/124732